pyproject.toml
[tool.poetry]
name = "AnimeRandomazer"
version = "0.0.1"
description = ""
authors = ["OSx"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyTelegramBotAPI = "^4.12.0"
beautifulsoup4 = "^4.12.2"
bs4 = "^0.0.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"